Fix SAA iterator max access
The bug was: When constructing an iterator using begin while the SAA was empty,
the ooa flag would not be initialized thus resulting in a non-equal-to end
value iterator resulting in a out of bound memory access.

Fixed by adding a checker for eventual out of bound initialization at the begin
iterator as it is equivalent to saying that this iterator already at end.